# sso/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  cotton-id_internal_httpx.Problem:
    properties:
      detail:
        description: Detail is a human-readable explanation specific to this occurrence.
        example: email is required
        type: string
      field:
        description: Field optionally names the offending request field for validation
          errors.
        example: email
        type: string
      instance:
        description: Instance optionally identifies the specific occurrence (the request
          path).
        type: string
      status:
        description: Status is the HTTP status code.
        example: 400
        type: integer
      title:
        description: Title is a short, human-readable summary of the problem type.
        example: Bad Request
        type: string
      type:
        description: |-
          Type is a URI reference identifying the problem type. "about:blank" is
          used when only the status conveys meaning.
        example: about:blank
        type: string
    type: object
...
  internal_auth.PublicUser:
    properties:
      about:
        type: string
      displayName:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: string
      location:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  internal_auth.forgotRequest:
    properties:
      email:
        example: alex@cotton-id.io
        type: string
    type: object
  internal_auth.loginRequest:
    properties:
      email:
        example: alex@cotton-id.io
        type: string
      password:
        example: hunter2Pass!
        type: string
      remember:
        example: true
        type: boolean
    type: object
  internal_auth.messageResponse:
    properties:
      message:
        example: If that email is registered, a reset link has been sent.
        type: string
    type: object
  internal_auth.resetRequest:
    properties:
      password:
        type: string
      token:
        type: string
    type: object
  internal_auth.signupRequest:
    properties:
      displayName:
        example: Alex Cotton
        type: string
      email:
        example: alex@cotton-id.io
        type: string
      password:
        example: hunter2Pass!
        type: string
      username:
        example: alex
        type: string
    type: object
  internal_auth.userEnvelope:
    properties:
      user:
        $ref: '#/definitions/internal_auth.PublicUser'
    type: object
...
info:
  contact: {}
  description: 'cotton-id is a single-tenant OpenID Connect identity provider. This
    is the change-1 walking skeleton: email/password auth, sessions, and the OIDC
    login/consent surface (OIDC and admin-client routes are stubs in this change).'
  title: cotton-id API
  version: 0.1.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates by email and password and establishes a session.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.userEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes the current session and clears the session cookie.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Log out
      tags:
      - auth
...
  /auth/password/forgot:
    post:
      consumes:
      - application/json
      description: Always responds 202 with a generic message; never reveals whether
        the email is registered.
      parameters:
      - description: Email to reset
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.forgotRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_auth.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Request a password reset
      tags:
      - auth
  /auth/password/reset:
    post:
      consumes:
      - application/json
      description: Sets a new password using a single-use, time-limited reset token;
        invalidates existing sessions.
      parameters:
      - description: Reset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.resetRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Reset a password
      tags:
      - auth
  /auth/session:
    get:
      description: Returns the authenticated user for the current session, or 401.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.userEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Current session
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new account with display name, username, email, and password,
        then establishes a session.
      parameters:
      - description: Signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.userEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/cotton-id_internal_httpx.Problem'
      summary: Create an account
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  AdminKey:
    in: header
    name: X-Admin-Key
    type: apiKey
  CSRF:
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"